Just some minor touchups; redid the enemy class which currently contains a default HP (to be set to 1, making use of Unreal's built in damage system). Got the flow of the player taking or dealing damage, doesn't work yet but just needs some more hammering.
Was looking into casting rays from the left and right hand side to detect whether the player was on a slope, but unreal crashed so the changes aren't there - still remember the gist of how i was setting it up.

Added a basic slope for testing, 40 degree angles will be the standard.